move save data use letter
